wire through `omit_load_extension` feature
this is a flag you can specify when compiling SQLite. When `omit_load_extension` is specified it changes the location of the sqlite routines from being installed in `papi_routines` to being in the global namespace.

ext.h has macros that flip based on the existed of `omit_load_extension` -- this update is now doing the same for rust. Granted we haven't put in our feature checks everywhere yet.